+++ CHANGELOG.md
@@ -5,7 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.0-rc.2] - UNRELEASED
+## [1.0.0] - 2020.06.02
+
+### Added
+### Changed / Improved
+
+- Support theme configuration via CLI (#369)
+- Update types for `getBundleOptionsValues` (#371)
+
+## [1.0.0-rc.2] - 2020-05-13
 
 ### Added